requirements.txt
streamlit
python-dotenv
langchain
langchain-community
openai
qrcode
faiss-cpu
PyPDF2
python-multipart
pillow
pandas
numpy